Fix worker proc hanging at exception
Not all exceptions in Task.run will lead to logging_thread.stop(),
which is required that the proc does not hang at exit,
because logging_thread was not a daemon thread.

We make it a daemon thread now.
And also, we additionally make extra sure that logging_thread.stop()
is called.
(Although there are maybe still other rare cases where this is not effective,
so the daemon thread property is still important as well.)

Fix #171.