Log finished outputs
Fix #152.

I used `FINISHED_LOG` as the name for the option
to have it consistent with `MEMORY_PROFILE_LOG`.

I used `"log/finished.log"` as the default path
to have it consistent with `"log/manager.log.%s"`
from the manager UI.

I thought about whether to have this
enabled or disabled by default,
and decided that this should not add too much overhead,
so we can have this enabled by default.